fix: fixed format type in components e2e

+++ test/e2e/components/components.test.js
@@ -46,7 +46,7 @@ test('should parse components jsdoc from jsdoc-example', async () => {
             year: {
               description: 'The year',
               type: 'number',
-              format: 'int64',
+              format: 'double',
             },
           },
         },
@@ -64,7 +64,7 @@ test('should parse components jsdoc from jsdoc-example', async () => {
             age: {
               description: 'Author age',
               type: 'number',
-              format: 'int64',
+              format: 'double',
             },
           },
         },
